Changes in the pieces handling and early implementation of Gravity
A major change was implemented in the the way that the game handles pieces. Instead of placing the piece directly into matrix, the pieces will only be placed in the matrix when they will be locked into place. This was changed due to the problem that the other method causes. It will not be possible to rotate or move the piece as it will cause collision with the piece itself when rotated or moved.

The other functions were changed to accomodate the change and an early implementation of the game "gravity" is implemented although laggy, which will require finding other solutions.

The piece controls were moved to the new function game_drop_piece and a function game_lock_piece writes the final position of the piece when reaches the bottom or touches another piece.