Addition of Features and Overall Adjustments
To improve code reability, comments describing functinons were added and further documentation will be created soon. Concerning the game mechanics, there is the following changes and additions:
 - Allowing the user to pause and resume the game
 - Reimplementation of the wall kicks, as previously they were programmed only to work with one single piece. Now, the wall kick algorithm tries to shift the pieces to the left or right, and if it is not possible, the piece will not rotate
 - The games' scores and dates are saved and will be display in the leaderboard
 - Testing regarding memory leaks were conducted and the program should not contain any
 - Minor Adjustments

At the moment, the main problem is to try to improve the keyboard input as it is possible to the user to hold the key down and the piece stays rotating or stuck at the place. Solutions were searched but no concrete idea and implementation was found