Improvements in Window Management and Minor Adjustments
Now, if the screen is too small, the program will ask the user to make the screen bigger and it does not crash if there not enough space to generate the game screen. The game menu can be resized in order to stay on center.

The last phase of development is the leaderboard, where minor implementations were made. After completed, the game will be finished and the documentations will be created accordingly